+++ app/features/ledger/SignCreatePltUpdate.ts
@@ -32,8 +32,9 @@ export default async function signUpdateCreatePltTransaction(
     let p1 = 0x00;
     const p2 = 0x00;
 
-    // Send initial packet of data
-    // 0x48	0x00 0x00 path_length path[uint32]x[5] update_instruction_header[28 bytes] update_type[uint8] payload_length[uint64] - Update type must be 24. Contains derivation path and transaction header.
+    // Send initial packet of data containing derivation path and transaction header.
+    // Update type must be 24.
+    // 0x48	0x00 0x00 path_length path[uint32]x[5] update_instruction_header[28 bytes] update_type[uint8] payload_length[uint64]
     const initialData = Buffer.concat([
         pathAsBuffer(path),
         serializedHeader,
@@ -42,8 +43,8 @@ export default async function signUpdateCreatePltTransaction(
     ]);
     await transport.send(0xe0, INS_PROTOCOL_UPDATE, p1, p2, initialData);
 
-    // Send message length.
-    // 0x48	0x01 0x00 token_symbol_length[uint32] [token_symbol[token_symbol_length bytes]] [token_module[32 bytes]] [decimals[uint8]] [initialization_params_length[uint32]] - Transaction payload containing token details and initialization parameters length. All fields are included in this command.
+    // Send part1 packet of data containing token details and initialization parameters length.
+    // 0x48	0x01 0x00 token_symbol_length[uint32] [token_symbol[token_symbol_length bytes]] [token_module[32 bytes]] [decimals[uint8]] [initialization_params_length[uint32]]
     p1 = 0x01;
     await transport.send(
         0xe0,
@@ -53,14 +54,16 @@ export default async function signUpdateCreatePltTransaction(
         serializedCreatePltUpdate.part1Buf
     );
 
-    // Stream the message bytes (maximum of 255 bytes per packet)
-    // 0x48	0x02 0x00 initialization_params[1...255 bytes] - Initialization parameters bytes. Sent in batches until the entirety of the initialization parameters (initialization_params_length bytes) has been sent. This command is repeated until all initialization parameter data has been sent.
+    // Stream the initParameter bytes (maximum of 255 bytes per packet).
+    // It is sent in batches until the entirety of the initialization parameters (initialization_params_length bytes) has been sent.
+    // 0x48	0x02 0x00 initialization_params[1...255 bytes]
     p1 = 0x02;
     const initParamChunks = chunkBuffer(
         serializedCreatePltUpdate.initParamBuf,
         255
     );
     for (let i = 0; i < initParamChunks.length; i += 1) {
+        // This command is repeated until all initialization parameter data has been sent.
         const result = await transport.send(
             0xe0,
             INS_PROTOCOL_UPDATE,